+++ src/main/java/edu/jhu/prim/sort/IntIntSort.java
@@ -16,9 +16,9 @@ public class IntIntSort {
      * are mirrored in index. Sorts in descending order.
      */
     public static void sortValuesDesc(int[] values, int[] index) {
-        IntArrays.scale(values, -1);
+        IntArrays.scale(values, (int) -1);
         sortValuesAsc(values, index);
-        IntArrays.scale(values, -1);
+        IntArrays.scale(values, (int) -1);
     }
     
     /**
@@ -68,9 +68,9 @@ public class IntIntSort {
      * are mirrored in values. Sorts in descending order.
      */
     public static void sortIndexDesc(int[] index, int[] values) {
-        IntArrays.scale(index, -1);
+        IntArrays.scale(index, (int) -1);
         sortIndexAsc(index, values);
-        IntArrays.scale(index, -1);
+        IntArrays.scale(index, (int) -1);
     }
     
     /**
@@ -159,7 +159,8 @@ public class IntIntSort {
     public static int[] getIntIndexArray(int length) {
         int[] index = new int[length];
         for (int i=0; i<index.length; i++) {
-            index[i] = i;
+            // TODO: This should maybe be a safe cast for the benefit of non-IntInt classes.
+            index[i] = (int) i;
         }
         return index;
     }
